[REF] point_of_sale: reorganize `utils` files structure
Reorganization of all utils files into a specific folder.

task-id: 4290514
