+++ addons/point_of_sale/static/src/app/screens/product_screen/control_buttons/control_buttons.js
@@ -5,7 +5,7 @@ import { SelectionPopup } from "@point_of_sale/app/components/popups/selection_p
 import { NoteButton } from "@point_of_sale/app/screens/product_screen/control_buttons/note_button/note_button";
 import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { _t } from "@web/core/l10n/translation";
-import { makeAwaitable } from "@point_of_sale/app/store/make_awaitable_dialog";
+import { makeAwaitable } from "@point_of_sale/app/utils/make_awaitable_dialog";
 import { SelectPartnerButton } from "@point_of_sale/app/screens/product_screen/control_buttons/select_partner_button/select_partner_button";
 
 export class ControlButtons extends Component {
