+++ addons/point_of_sale/static/src/app/screens/ticket_screen/invoice_button/invoice_button.js
@@ -3,7 +3,7 @@ import { useService } from "@web/core/utils/hooks";
 import { AlertDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { Component } from "@odoo/owl";
-import { ask, makeAwaitable } from "@point_of_sale/app/store/make_awaitable_dialog";
+import { ask, makeAwaitable } from "@point_of_sale/app/utils/make_awaitable_dialog";
 import { PartnerList } from "../../partner_list/partner_list";
 
 export class InvoiceButton extends Component {
