+++ addons/point_of_sale/static/src/app/utils/debug/debug_widget.xml
