+++ addons/point_of_sale/static/src/app/utils/error_handlers.js
