+++ addons/point_of_sale/static/src/app/utils/make_awaitable_dialog.js
