+++ addons/point_of_sale/static/src/app/utils/printer/base_printer.js
