+++ addons/pos_event/static/src/app/screens/product_screen/product_screen.js
@@ -1,5 +1,5 @@
 import { ProductScreen } from "@point_of_sale/app/screens/product_screen/product_screen";
-import { makeAwaitable } from "@point_of_sale/app/store/make_awaitable_dialog";
+import { makeAwaitable } from "@point_of_sale/app/utils/make_awaitable_dialog";
 import { patch } from "@web/core/utils/patch";
 import { EventConfiguratorPopup } from "@pos_event/app/popup/event_configurator_popup/event_configurator_popup";
 import { _t } from "@web/core/l10n/translation";
