+++ addons/pos_mercado_pago/static/src/app/payment_mercado_pago.js
@@ -1,5 +1,5 @@
 import { _t } from "@web/core/l10n/translation";
-import { PaymentInterface } from "@point_of_sale/app/payment/payment_interface";
+import { PaymentInterface } from "@point_of_sale/app/utils/payment/payment_interface";
 import { AlertDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 
 export class PaymentMercadoPago extends PaymentInterface {
