+++ addons/pos_restaurant/static/src/overrides/models/payment.js
@@ -1,4 +1,4 @@
-import { PaymentInterface } from "@point_of_sale/app/payment/payment_interface";
+import { PaymentInterface } from "@point_of_sale/app/utils/payment/payment_interface";
 import { patch } from "@web/core/utils/patch";
 
 patch(PaymentInterface.prototype, {
