+++ addons/pos_self_order/__manifest__.py
@@ -48,7 +48,6 @@
             'web/static/lib/bootstrap/js/dist/base-component.js',
             "web/static/lib/bootstrap/js/dist/carousel.js",
             'web/static/lib/bootstrap/js/dist/scrollspy.js',
-            "point_of_sale/static/src/app/store/models/product_custom_attribute.js",
             'web_editor/static/src/js/editor/odoo-editor/src/base_style.scss',
             'web_editor/static/src/scss/web_editor.common.scss',
             "point_of_sale/static/src/app/generic_components/numpad/*",
@@ -59,12 +58,12 @@
             "point_of_sale/static/src/css/pos_receipts.css",
             "point_of_sale/static/src/app/screens/receipt_screen/receipt/**/*",
             "pos_self_order/static/src/overrides/components/receipt_header/*",
-            "point_of_sale/static/src/app/printer/base_printer.js",
+            "point_of_sale/static/src/app/utils/printer/base_printer.js",
             "point_of_sale/static/src/app/services/printer_service.js",
             'point_of_sale/static/src/app/utils/html-to-image.js',
             "point_of_sale/static/src/app/services/render_service.js",
             "pos_self_order/static/src/app/**/*",
-            "point_of_sale/static/src/app/printer/hw_printer.js",
+            "point_of_sale/static/src/app/utils/printer/hw_printer.js",
             "web/static/src/core/utils/render.js",
             "pos_self_order/static/src/app/store/order_change_receipt_template.xml",
             "account/static/src/helpers/*.js",
