+++ addons/pos_stripe/static/src/app/payment_stripe.js
@@ -1,7 +1,7 @@
 /* global StripeTerminal */
 
 import { _t } from "@web/core/l10n/translation";
-import { PaymentInterface } from "@point_of_sale/app/payment/payment_interface";
+import { PaymentInterface } from "@point_of_sale/app/utils/payment/payment_interface";
 import { AlertDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 
 export class PaymentStripe extends PaymentInterface {
